[RESTEASY-2024] MediaTypeEditor to a package compliant with java module rules. Relevant documentation added to class's javadoc
